requirements.txt
# StockTradebyZ-multimodal 全流程依赖
# 基础（必装）
PyYAML>=6.0
pandas>=1.3.0
numpy>=1.20.0
tqdm>=4.60.0

# 步骤 1 拉取 K 线（需 TUSHARE_TOKEN）
tushare>=1.2.0

# 步骤 2 量化初选（Numba 加速）
numba>=0.55.0

# 步骤 3 导出 K 线图（Plotly + 静态图导出）
plotly>=5.0.0
kaleido>=0.2.0

# 多模态复评（按 config/review.yaml 中的 provider 任选其一安装）
# Gemini
google-genai>=1.0.0
# OpenAI / 通义 / 智谱 / 月之暗面 / DeepSeek（兼容 OpenAI 接口）
openai>=1.0.0
# Claude
anthropic>=0.39.0
